Improve error message formatting
slack error responses should have all the error details in the error field -
that's not what's happening in reality though, e.g.

{
  "ok":false,
   "error":"missing_scope",
   "needed":"emoji:read",
   "provided":"identify,bot"
}

I can't think of a generic but pretty way of handling this so let's go with
generic but ugly: just add the original json response to the error message as
well so no details from other fields go missing.